Implement reverse image order in gallery view, closes #92
This update adds a toggle for reverse image order in the PictureList component. The user’s preferred order is stored in the browser for future visits. This feature has been integrated into the ReceptionPictures and NavContentPictures components. The picture order can now be controlled in the UI, providing a more customizable user experience.